fix(select): restore test categorization and contracts
